move logger.py to handle complete package, not just script-related files
